Checkpoint: Ripping apart bookkeeper
