+++ bout_runners/bookkeeper/bookkeeper_utils.py
@@ -144,134 +144,6 @@ def get_system_info_as_sql_type():
     return sys_info_dict
 
 
-def get_create_table_statement(table_name,
-                               columns=None,
-                               primary_key='id',
-                               foreign_keys=None):
-    """
-    Return a SQL string which can be used to create the table.
-
-    Parameters
-    ----------
-    table_name : str
-        Name of the table
-    columns : dict or None
-        Dictionary where the key is the column name and the value is
-        the type
-    primary_key : str
-        Name of the primary key (the type is set to INTEGER)
-    foreign_keys : dict or None
-        Dictionary where the key is the column in this table to be
-        used as a foreign key and the value is the tuple
-        consisting of (name_of_the_table, key_in_table) to refer to
-
-    Returns
-    -------
-    create_statement : str
-        The SQL statement which creates table
-    """
-    create_statement = f'CREATE TABLE {table_name} \n('
-
-    create_statement += f'   {primary_key} INTEGER PRIMARY KEY,\n'
-
-    # These are not known during submission time
-    nullable_fields = ('start_time',
-                       'stop_time')
-    if columns is not None:
-        for name, sql_type in columns.items():
-            create_statement += f'    {name} {sql_type}'
-            if name in nullable_fields:
-                nullable_str = ',\n'
-            else:
-                nullable_str = ' NOT NULL,\n'
-            create_statement += nullable_str
-
-    if foreign_keys is not None:
-        # Create the key as column
-        # NOTE: All keys are integers
-        for name in foreign_keys.keys():
-            create_statement += f'    {name} INTEGER NOT NULL,\n'
-
-        # Add constraint
-        for name, (f_table_name, key_in_table) in foreign_keys.items():
-            # If the parent is updated or deleted, we would like the
-            # same effect to apply to its child (thereby the CASCADE
-            # parameter)
-            create_statement += \
-                (f'    FOREIGN KEY({name}) \n'
-                 f'        REFERENCES {f_table_name}({key_in_table})\n'
-                 f'            ON UPDATE CASCADE\n'
-                 f'            ON DELETE CASCADE,'
-                 f'\n')
-
-    # Replace last comma with )
-    create_statement = f'{create_statement[:-2]})'
-
-    return create_statement
-
-
-def get_db_path(database_root_path, project_path):
-    """
-    Return the database path.
-
-    Parameters
-    ----------
-    project_path : None or Path or str
-        Root path to the project (i.e. where the makefile is located)
-        If None the root caller directory will be used
-    database_root_path : None or Path or str
-        Root path of the database file
-        If None, the path will be set to $HOME/BOUT_db
-
-    Returns
-    -------
-    database_path : Path
-        Path to the database
-    """
-    if project_path is None:
-        project_path = get_caller_dir()
-
-    if database_root_path is None:
-        # NOTE: This will be changed when going to production
-        database_root_path = get_caller_dir()
-        # database_root_path = Path().home().joinpath('BOUT_db')
-
-    project_path = Path(project_path)
-    database_root_path = Path(database_root_path)
-
-    database_root_path.mkdir(exist_ok=True, parents=True)
-    # NOTE: sqlite does not support schemas (except through an
-    #       ephemeral ATTACH connection)
-    #       Thus we will make one database per project
-    # https://www.sqlite.org/lang_attach.html
-    # https://stackoverflow.com/questions/30897377/python-sqlite3-create-a-schema-without-having-to-use-a-second-database
-    schema = project_path.name
-    database_path = database_root_path.joinpath(f'{schema}.db')
-
-    return database_path
-
-
-def tables_created(bookkeeper):
-    """
-    Check if the tables is created in the database.
-
-    Parameters
-    ----------
-    bookkeeper : Bookkeeper
-        The Bookkeeper object
-
-    Returns
-    -------
-    bool
-        Whether or not the tables are created
-    """
-    query_str = ('SELECT name FROM sqlite_master '
-                 '   WHERE type="table"')
-
-    table = bookkeeper.query(query_str)
-    return len(table.index) != 0
-
-
 def extract_parameters_in_use(project_path,
                               bout_inp_dst_dir,
                               run_parameters_dict):
@@ -310,34 +182,6 @@ def extract_parameters_in_use(project_path,
     parameters.update(run_parameters_dict)
 
     return parameters
-
-
-def create_insert_string(field_names, table_name):
-    """
-    Create a question mark style string for database insertions.
-
-    Values must be provided separately in the execution statement
-
-    Parameters
-    ----------
-    field_names : array-like
-        Names of the fields to populate
-    table_name : str
-        Name of the table to use for the insertion
-
-    Returns
-    -------
-    insert_str : str
-        The string to be used for insertion
-    """
-    # From
-    # https://stackoverflow.com/a/14108554/2786884
-    columns = ', '.join(field_names)
-    placeholders = ', '.join('?' * len(field_names))
-    insert_str = f'INSERT INTO {table_name} ' \
-                 f'({columns}) ' \
-                 f'VALUES ({placeholders})'
-    return insert_str
 
 
 def get_file_modification(project_path, makefile_path, exec_name):
